fix dupe route-data
