+++ public/canjs/routing/route-data/route-data.js
@@ -36,8 +36,20 @@ import {
   toSelectedParts,
 } from "../data-utils.js";
 
-import { getTimingLevels } from "../../../react/SettingsSidebar/components/TimingCalculation/helpers";
+export function getTimingLevels(issueHierarchy, timingCalculations) {
+  const baseLevels = createBaseLevels(issueHierarchy);
 
+  return baseLevels.map((level, i) => {
+    const child = baseLevels[i + 1];
+    const isLast = i === baseLevels.length - 1;
+
+    return {
+      ...level,
+      childType: child ? child.type : null,
+      calculations: calculationsForLevel(level, child, timingCalculations[level.type], isLast),
+    };
+  });
+}
 const _15DAYS_IN_S = (DAY_IN_MS / 1000) * 15;
 
 const booleanParsing = {
